move `ExpandProcAttrMacros` to `lib.rs`
as `db.rs` is going to be removed
