+++ src/tools/rust-analyzer/crates/hir-def/src/nameres/collector.rs
@@ -617,7 +617,7 @@ impl<'db> DefCollector<'db> {
         let (expander, kind) = match self.proc_macros.iter().find(|(n, _, _)| n == &def.name) {
             Some(_)
                 if kind == hir_expand::proc_macro::ProcMacroKind::Attr
-                    && !crate::db::ExpandProcAttrMacros::get(self.db).enabled(self.db) =>
+                    && !crate::ExpandProcAttrMacros::get(self.db).enabled(self.db) =>
             {
                 (CustomProcMacroExpander::disabled_proc_attr(), kind)
             }
