+++ src/tools/rust-analyzer/crates/hir/src/db.rs
@@ -3,8 +3,5 @@
 //! be cool if we didn't do that.
 //!
 //! But we need this for at least LRU caching at the query level.
-pub use hir_def::{
-    db::{DefDatabase, set_expand_proc_attr_macros},
-    file_item_tree,
-};
+pub use hir_def::{db::DefDatabase, file_item_tree, set_expand_proc_attr_macros};
 pub use hir_ty::db::HirDatabase;
